server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
ollama==0.6.1
pydantic==2.12.5
# emoji==2.10.0
# numpy
